Issue #56: Refresh player and retry URL decoding on 403
Introduce a reattempt in playback._get_track on
encountering a 403 error from the verify track URL.
Refresh the youtube player and re-invoke the _get_track
method so that the URL is decoded again.
Only one reattempt is allowed.

Also introduces a wait parameter on the RepeatingTimer
now() method to block until the update method has run.

Initialize youtube player on startup (before timer)